LayoutTests/fast/css/first-letter-text-transform-change-expected.txt

ab
AB
PASS: d1 text content is 'ab'
PASS: d2 text content is 'AB'
PASS: d1 has both characters (width=40)
PASS: d2 has both characters (width=40)
